+++ lib/index.js
@@ -1,6 +1,6 @@
 const logger = require('./logger');
 const configParser = require('./config');
-const repoWorker = require('./workers/repo');
+const repositoryWorker = require('./workers/repository');
 
 module.exports = {
   start,
@@ -14,7 +14,7 @@ async function start() {
     // Iterate through repositories sequentially
     for (const repository of config.repositories) {
       const repoConfig = getRepoConfig(config, repository);
-      await repoWorker.processRepo(repoConfig, logger);
+      await repositoryWorker.processRepo(repoConfig, logger);
     }
     logger.info('Renovate finished');
   } catch (error) {
